rename: components/ui index 관리(#206)
