+++ src/features/common/CustomBarChart.tsx
@@ -1,4 +1,4 @@
-import { ChartContainer } from '@/components/ui/chart';
+import { ChartContainer } from '@/components/ui';
 import { useEffect, useState } from 'react';
 import { Bar, BarChart, CartesianGrid, XAxis, YAxis } from 'recharts';
 import { RecentRecord, YearRecord } from '../player/types/detail';
