+++ src/features/player/components/detail/StatCard.tsx
@@ -1,4 +1,4 @@
-import { Progress } from '@/components/ui/progress';
+import { Progress } from '@/components/ui';
 
 interface StatCardProps {
   header: string;
